GameType type should reside in features/gameTypes
