+++ src/app/features/gameTypes/GameType.ts
@@ -1,0 +1,12 @@
+import type { RuntimeDefinition } from "../compiler/types";
+import type {
+  GameDefinition,
+  GameTypeId,
+} from "../../../api/services/game/types";
+
+export interface GameType {
+  id: GameTypeId;
+  name: string;
+  defaultGameDefinition: GameDefinition;
+  runtimeDefinition: RuntimeDefinition;
+}
