+++ src/app/features/gameTypes/index.ts
@@ -1,6 +1,7 @@
-import type { GameType, GameTypeId } from "../../../api/services/game/types";
+import type { GameTypeId } from "../../../api/services/game/types";
 import { reactVersus } from "./versus/react/gameType";
 import { pixiVersus } from "./versus/pixi/gameType";
+import type { GameType } from "./GameType";
 
 export const gameTypeList: GameType[] = [reactVersus, pixiVersus];
 export const gameTypes = gameTypeList.reduce(
