+++ src/app/features/gameTypes/versus/pixi/gameType.ts
@@ -1,10 +1,8 @@
-import type {
-  GameType,
-  GameTypeId,
-} from "../../../../../api/services/game/types";
+import type { GameTypeId } from "../../../../../api/services/game/types";
 import { gameDefinitionType } from "../../../../../api/services/game/types";
 import defaultGameDefinitionJson from "../defaultGameDefinition.json";
 import { reactVersusDefinition } from "../runtimeDefinition";
+import type { GameType } from "../../GameType";
 
 export const pixiVersus: GameType = {
   id: "pixi-versus" as GameTypeId,
