+++ src/app/features/gameTypes/versus/react/gameType.ts
@@ -1,10 +1,8 @@
-import type {
-  GameType,
-  GameTypeId,
-} from "../../../../../api/services/game/types";
+import type { GameTypeId } from "../../../../../api/services/game/types";
 import { reactVersusDefinition } from "../runtimeDefinition";
 import { gameDefinitionType } from "../../../../../api/services/game/types";
 import defaultGameDefinitionJson from "../defaultGameDefinition.json";
+import type { GameType } from "../../GameType";
 
 export const reactVersus: GameType = {
   id: "react-versus" as GameTypeId,
